async/await

+++ controllers/feed.js
@@ -3,30 +3,28 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page;
     const perPage = 2;
-    let totalItems;
-    Post.find().countDocuments()
-        .then(count => {
-            totalItems = count;
-            return Post.find().skip((currentPage - 1) * perPage).limit(perPage)
+    try {
+        const totalItems = await Post.find().countDocuments()
+        const posts = await Post.find().skip((currentPage - 1) * perPage).limit(perPage)
+
+        res.status(200).json({
+            message: 'Fetched posts successfully',
+            posts: posts,
+            totalItems: totalItems
         })
-        .then(posts => {
-            res.status(200).json({
-                message: 'Fetched posts successfully',
-                posts: posts,
-                totalItems: totalItems
-            })
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
-        })
-};
+    }
+    catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    };
+}
 
 exports.createPost = (req, res, next) => {
     const errors = validationResult(req);
@@ -43,26 +41,35 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: "Trinh"
-        },
+        creator: req.userId,
     })
 
     post.save().then(result => {
-        res.status(201).json({
-            message: 'Post created successfully!',
-            post: result
-        });
-    }).catch(err => {
-        if (!err.statusCode) {
-            err.statusCode = 500;
-        }
-        next(err);
+        return User.findById(req.userId);
     })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
+            res.status(201).json({
+                message: 'Post created successfully!',
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
+            });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
+        })
 };
 
 exports.getPost = (req, res, next) => {
@@ -111,6 +118,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized.')
+            error.statusCode = 403;
+            throw error;
+        }
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -142,8 +154,21 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.')
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             console.log(result);
